refactor: extract manager product methods into dedicated service
